+++ packages/hydrogen/CHANGELOG.md
@@ -28,6 +28,7 @@ and adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
   - `useCartLinesUpdateCallback`
   - `useCartNoteUpdateCallback`
 - Feat: Transition away from deprecated currency selector in favor of country selector
+- Fix: add charset to content type in HTML responses
 
 ## 0.10.1 - 2022-01-26